requirements.txt
psutil
rich
Pillow